src/backend/requirementsDocker.txt
fastapi
gritlm
huggingface_hub
numba
numpy
openai
openai-whisper
pyannote.audio
pyannote.core
pyannote.database
pyannote.metrics
pyannote.pipeline
pydantic
python-dotenv
scipy
sentence_transformers
starlette
tiktoken
torch
tqdm
transformers
weaviate_client
